Move tests into `assertions` and call from an overall test file.
